+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/types/ImmutableAbstractMessage.java
@@ -11,7 +11,8 @@ import java.util.EnumSet;
  * @param <T> The type of the immutable type
  * @param <S> The type of the packet form of this class
  */
-public abstract class ImmutableAbstractMessage<T extends ImmutableAbstractMessage<T, S>, S> {
+sealed public abstract class ImmutableAbstractMessage<T extends ImmutableAbstractMessage<T, S>, S>
+		permits ImmutableC2SChatMessage, ImmutableS2CChatMessage, ImmutableS2CSystemMessage, MutableAbstractMessage {
 	protected final @NotNull PlayerEntity player;
 	protected final boolean isOnClientSide;
