+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/types/ImmutableC2SChatMessage.java
@@ -1,4 +1,4 @@
-package org.quiltmc.qsl.chat.api.types.c2s;
+package org.quiltmc.qsl.chat.api.types;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.message.MessageSignature;
@@ -7,19 +7,18 @@ import net.minecraft.network.packet.c2s.play.ChatMessageC2SPacket;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.quiltmc.qsl.chat.api.QuiltMessageType;
-import org.quiltmc.qsl.chat.api.types.MutableAbstractMessage;
 
 import java.time.Instant;
 import java.util.EnumSet;
 
-public class MutableC2SChatMessage extends MutableAbstractMessage<ImmutableC2SChatMessage, ChatMessageC2SPacket> {
-	private String message;
-	private Instant timestamp;
-	private long salt;
-	private @Nullable MessageSignature signature;
-	private MessageSignatureList.Acknowledgment messageAcknowledgments;
+public final class ImmutableC2SChatMessage extends ImmutableAbstractMessage<ImmutableC2SChatMessage, ChatMessageC2SPacket> {
+	private final String message;
+	private final Instant timestamp;
+	private final long salt;
+	private final @Nullable MessageSignature signature;
+	private final MessageSignatureList.Acknowledgment messageAcknowledgments;
 
-	public MutableC2SChatMessage(PlayerEntity player, ChatMessageC2SPacket packet) {
+	public ImmutableC2SChatMessage(PlayerEntity player, ChatMessageC2SPacket packet) {
 		this(
 				player,
 				packet.message(),
@@ -29,7 +28,8 @@ public class MutableC2SChatMessage extends MutableAbstractMessage<ImmutableC2SCh
 				packet.messageAcknowledgments()
 		);
 	}
-	public MutableC2SChatMessage(PlayerEntity player, String message, Instant timestamp, long salt, @Nullable MessageSignature signature, MessageSignatureList.Acknowledgment messageAcknowledgments) {
+
+	public ImmutableC2SChatMessage(PlayerEntity player, String message, Instant timestamp, long salt, @Nullable MessageSignature signature, MessageSignatureList.Acknowledgment messageAcknowledgments) {
 		super(player);
 		this.message = message;
 		this.timestamp = timestamp;
@@ -57,39 +57,19 @@ public class MutableC2SChatMessage extends MutableAbstractMessage<ImmutableC2SCh
 		return message;
 	}
 
-	public void setMessage(String message) {
-		this.message = message;
-	}
-
 	public Instant getTimestamp() {
 		return timestamp;
-	}
-
-	public void setTimestamp(Instant timestamp) {
-		this.timestamp = timestamp;
 	}
 
 	public long getSalt() {
 		return salt;
 	}
 
-	public void setSalt(long salt) {
-		this.salt = salt;
-	}
-
 	public @Nullable MessageSignature getSignature() {
 		return signature;
 	}
 
-	public void setSignature(@Nullable MessageSignature signature) {
-		this.signature = signature;
-	}
-
 	public MessageSignatureList.Acknowledgment getMessageAcknowledgments() {
 		return messageAcknowledgments;
-	}
-
-	public void setMessageAcknowledgments(MessageSignatureList.Acknowledgment messageAcknowledgments) {
-		this.messageAcknowledgments = messageAcknowledgments;
 	}
 }
