+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/types/MutableAbstractMessage.java
@@ -2,7 +2,8 @@ package org.quiltmc.qsl.chat.api.types;
 
 import net.minecraft.entity.player.PlayerEntity;
 
-abstract public class MutableAbstractMessage<T extends ImmutableAbstractMessage<T, S>, S> extends ImmutableAbstractMessage<T, S> {
+sealed abstract public class MutableAbstractMessage<T extends ImmutableAbstractMessage<T, S>, S> extends ImmutableAbstractMessage<T, S>
+		permits ImmutableS2CProfileIndependentMessage, MutableC2SChatMessage, MutableS2CChatMessage, MutableS2CProfileIndependentMessage, MutableS2CSystemMessage {
 	public MutableAbstractMessage(PlayerEntity player, boolean isOnClientSide) {
 		super(player, isOnClientSide);
 	}
