+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/api/types/MutableS2CSystemMessage.java
@@ -1,15 +1,14 @@
-package org.quiltmc.qsl.chat.api.types.s2c;
+package org.quiltmc.qsl.chat.api.types;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.packet.s2c.play.SystemMessageS2CPacket;
 import net.minecraft.text.Text;
 import org.jetbrains.annotations.NotNull;
 import org.quiltmc.qsl.chat.api.QuiltMessageType;
-import org.quiltmc.qsl.chat.api.types.MutableAbstractMessage;
 
 import java.util.EnumSet;
 
-public class MutableS2CSystemMessage extends MutableAbstractMessage<ImmutableS2CSystemMessage, SystemMessageS2CPacket> {
+public final class MutableS2CSystemMessage extends MutableAbstractMessage<ImmutableS2CSystemMessage, SystemMessageS2CPacket> {
 	private Text content;
 	private boolean overlay;
