+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/mixin/ServerPlayNetworkHandlerMixin.java
@@ -8,12 +8,12 @@ import net.minecraft.server.network.ServerPlayNetworkHandler;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
 import org.quiltmc.qsl.chat.api.QuiltChatEvents;
-import org.quiltmc.qsl.chat.api.types.c2s.ImmutableC2SChatMessage;
-import org.quiltmc.qsl.chat.api.types.s2c.ImmutableS2CChatMessage;
-import org.quiltmc.qsl.chat.api.types.s2c.ImmutableS2CProfileIndependentMessage;
-import org.quiltmc.qsl.chat.api.types.c2s.MutableC2SChatMessage;
-import org.quiltmc.qsl.chat.api.types.s2c.MutableS2CChatMessage;
-import org.quiltmc.qsl.chat.api.types.s2c.MutableS2CProfileIndependentMessage;
+import org.quiltmc.qsl.chat.api.types.ImmutableC2SChatMessage;
+import org.quiltmc.qsl.chat.api.types.ImmutableS2CChatMessage;
+import org.quiltmc.qsl.chat.api.types.ImmutableS2CProfileIndependentMessage;
+import org.quiltmc.qsl.chat.api.types.MutableC2SChatMessage;
+import org.quiltmc.qsl.chat.api.types.MutableS2CChatMessage;
+import org.quiltmc.qsl.chat.api.types.MutableS2CProfileIndependentMessage;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
