+++ library/management/chat/src/main/java/org/quiltmc/qsl/chat/mixin/client/ClientPlayNetworkHandlerMixin.java
@@ -7,7 +7,12 @@ import net.minecraft.network.packet.s2c.play.ProfileIndependentMessageS2CPacket;
 import net.minecraft.network.packet.s2c.play.SystemMessageS2CPacket;
 import org.objectweb.asm.Opcodes;
 import org.quiltmc.qsl.chat.api.QuiltChatEvents;
-import org.quiltmc.qsl.chat.api.types.s2c.*;
+import org.quiltmc.qsl.chat.api.types.ImmutableS2CChatMessage;
+import org.quiltmc.qsl.chat.api.types.ImmutableS2CProfileIndependentMessage;
+import org.quiltmc.qsl.chat.api.types.ImmutableS2CSystemMessage;
+import org.quiltmc.qsl.chat.api.types.MutableS2CChatMessage;
+import org.quiltmc.qsl.chat.api.types.MutableS2CProfileIndependentMessage;
+import org.quiltmc.qsl.chat.api.types.MutableS2CSystemMessage;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
